Moved getMyClasses function to user routes and user controller
