+++ src/routes/users_routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const usersRouter = express.Router();
-const {getAllUsers, login, updateUser, deleteUser, register, getUserByID} = require('../controllers/users_controller');
+const {getAllUsers, login, updateUser, deleteUser, register, getUserByID, getMyClasses} = require('../controllers/users_controller');
 const {validateRequest, validateAdmin} = require('../middlewares/auth_middleware');
 
 // Public routes (No authentication required)
@@ -10,6 +10,7 @@ usersRouter.post("/login", login);
 // Routes that require user authentication
 usersRouter.use(validateRequest);
 usersRouter.get("/myaccount", getUserByID);
+usersRouter.get("/myclasses", getMyClasses);
 usersRouter.put("/:id", updateUser);
 
 // Routes that require both user and admin authentication
